requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
discord==1.0.1
discord.py==1.3.2
idna==2.9
multidict==4.7.5
numpy==1.18.2
python-dotenv==0.12.0
requests==2.23.0
scipy==1.4.1
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
